// src/api/generated/models/genreCatalogResponse.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Manyak Server API
 * Manyak backend API documentation
 * OpenAPI spec version: v1
 */
import type { GenreCatalogItem } from './genreCatalogItem';

export interface GenreCatalogResponse {
  genres?: GenreCatalogItem[];
  featuredGenres?: GenreCatalogItem[];
}